Add files via upload
project uses the Turtle graphics library i started learning multiple days ago and created some games with, to create an interactive game where users guess the names of U.S. states. The game displays a blank map of the United States, and players input state names. Correctly guessed states are marked on the map. If the user types “Exit,” the game ends, and a CSV file (states_to_learn.csv) is generated with the remaining states to learn.